move multiple textstring
