+++ src/packages/property-editors/multiple-text-string/manifests.ts
