+++ src/packages/property-editors/multiple-text-string/property-editor-ui-multiple-text-string.element.ts
@@ -1,4 +1,4 @@
-import { UmbPropertyValueChangeEvent } from '../../index.js';
+import { UmbPropertyValueChangeEvent } from '../../core/property-editor/index.js';
 import type { UmbChangeEvent } from '@umbraco-cms/backoffice/event';
 import { html, customElement, property, state, ifDefined } from '@umbraco-cms/backoffice/external/lit';
 import type { UmbPropertyEditorConfigCollection } from '@umbraco-cms/backoffice/property-editor';
