+++ src/packages/property-editors/multiple-text-string/property-editor-ui-multiple-text-string.stories.ts
